+++ pages/update_data.php
@@ -758,7 +758,92 @@ if($_GET['ingSupplier'] == 'delete'){
 	return;
 }
 
-
+//FORMULA QUANTITY MANAGEMENT
+if($_POST['updateQuantity'] && $_POST['ingQuantityID'] &&  $_POST['ingQuantityName']  && $_POST['fid']){
+	$fid = $_POST['fid'];
+	$value = $_POST['ingQuantity'];
+	$ingredient = $_POST['ingQuantityID'];
+	$ing_name = $_POST['ingQuantityName'];
+	
+	if(empty($_POST['ingQuantity'])){
+		$response["error"] = 'Quantity cannot be empty.';
+		echo json_encode($response);
+		return;
+	}
+	if(!is_numeric($_POST['ingQuantity'])){
+		$response["error"] = 'Quantity must be numeric only.';
+		echo json_encode($response);
+		return;
+	}
+	
+	if($_POST['curQuantity'] == $_POST['ingQuantity']){
+		$response["error"] = 'Quantity is already the same.';
+		echo json_encode($response);
+		return;
+	}
+	
+	
+	
+	if($_POST['ingReCalc'] == 'true'){
+		$ingID = $_POST['ingID'];
+		
+		$q1 = mysqli_query($conn,"SELECT formulas.ingredient,formulas.ingredient_id,formulas.quantity,ingredients.profile FROM formulas,ingredients WHERE fid = '".$fid."' AND ingredients.id = formulas.ingredient_id AND ingredients.profile='solvent'");
+		while($r = mysqli_fetch_array($q1)){
+			
+			$slv[] = $r;
+			
+		}
+		
+		if(!$slv[0]['profile']){
+			$response["error"] = 'No solvent available';
+			echo json_encode($response);
+			return;
+		}
+		
+		if($slv[0]['quantity'] < $_POST['ingQuantity']){
+			$response["error"] = 'Not enough solvent, available: '.$slv[0]['quantity'];
+			echo json_encode($response);
+			return;
+		}
+		
+		//UPDATE SOLVENT
+		function formatVal($num){
+    		return sprintf("%+d",$num);
+		}
+		
+		$curV = mysqli_fetch_array(mysqli_query($conn, "SELECT quantity FROM formulas WHERE fid = '$fid' AND id = '".$ingredient."'"));
+		$diff = number_format(  $curV['quantity'] -  $value  , 4);
+		$v = formatVal($diff);
+		
+		$qs ="UPDATE formulas SET quantity = quantity $v WHERE fid = '$fid' AND ingredient_id = '".$slv[0]['ingredient_id']."'";
+		if(!mysqli_query($conn, $qs)){
+			$response["error"] = 'Error updating solvent: '.mysqli_error($conn);
+			$response["query"] = $qs;
+			echo json_encode($response);
+			return;
+		}
+	}
+		
+	$meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,isProtected FROM formulasMetaData WHERE fid = '".$_POST['fid']."'"));
+	
+	if($meta['isProtected'] == FALSE){
+		
+		if(mysqli_query($conn, "UPDATE formulas SET quantity = '$value' WHERE fid = '$fid' AND id = '$ingredient'")){
+			
+			
+			$lg = "CHANGE: ".$ing_name." Set $name to $value";
+			mysqli_query($conn, "INSERT INTO formula_history (fid,change_made,user) VALUES ('".$meta['id']."','$lg','".$user['fullName']."')");
+			$response["success"] = 'Quantity updated';
+			echo json_encode($response);
+		
+		}else{
+			$response["error"] = mysqli_error($conn);
+			echo json_encode($response);
+		}
+		
+	}
+	return;
+}
 
 if($_POST['value'] && $_GET['formula'] && $_POST['pk']){
 	$value = mysqli_real_escape_string($conn, $_POST['value']);