+++ releasenotes.md
@@ -16,4 +16,6 @@ Whats New in v7.4
 - Minor UI improvements in formula making view
 - Functionality to add comments/notes when adding ingredient in formula making view
 - Added a dropdown menu to view original formula from scheduled formulas
-
+- Added aditional info in the popup message when replacing an ingredient in formula
+- Change formula history order by most recent changes
+- Added advanced formula editor (quantity only)